align json and text output
