+++ packages/third-party-notices/src/output/json.ts
@@ -1,41 +1,5 @@
-import { findPackage, readPackage } from "@rnx-kit/tools-node/package";
 import type { License, LicenseJSONInfo } from "../types";
-
-function getPackageAuthor(modulePath: string): string | undefined {
-  const pkgFile = findPackage(modulePath);
-  if (pkgFile) {
-    const manifest = readPackage(pkgFile);
-    if (manifest) {
-      return typeof manifest.author === "string"
-        ? manifest.author
-        : manifest.author?.name;
-    }
-  }
-  return undefined;
-}
-
-function parseCopyright(
-  modulePath: string,
-  licenseText: string | undefined,
-  license: string | undefined,
-  licenseURLs: string[]
-): string {
-  const m = licenseText?.match(/^Copyright .*$/m);
-  if (!m) {
-    const packageAuthor = getPackageAuthor(modulePath);
-    if (packageAuthor) {
-      return packageAuthor;
-    }
-
-    if (licenseURLs?.length > 0) {
-      return `${license} (${licenseURLs.join(" ")})`;
-    }
-
-    return "No copyright notice";
-  }
-
-  return m[0].trim();
-}
+import { parseCopyright } from "./copyright";
 
 export function createLicenseJSON(
   licenses: License[],
@@ -53,7 +17,9 @@ export function createLicenseJSON(
           licenseURLs,
         }) => {
           if (!license) {
-            throw new Error(`No license for ${name}`);
+            throw new Error(
+              `No license information found for package '${name}'. Consider filing an issue for the project to properly advertise its licence. Pass this module to the tool via '--ignoreModules ${name}' to suppress this message.`
+            );
           }
           const info: LicenseJSONInfo = {
             name,
