+++ packages/third-party-notices/src/output/text.ts
@@ -1,4 +1,5 @@
 import type { License } from "../types";
+import { getPackageAuthor } from "./copyright";
 
 const EOL = "\n";
 const SEPARATOR = `${EOL}${EOL}========================================================================${EOL}${EOL}`;
@@ -11,24 +12,40 @@ export function createLicenseFileContents(
   const output = preambleText ? [preambleText.join(EOL)] : [];
 
   // Emit combined license text
-  licenses.forEach(({ name, version, license, licenseText, licenseURLs }) => {
-    if (license?.toUpperCase() === "UNLICENSED") {
-      // Ignore unlicensed/private packages
-      return;
-    }
-
-    if (!licenseText) {
-      if (!license && (!licenseURLs || licenseURLs.length === 0)) {
-        throw new Error(
-          `No license information found for package '${name}'. Consider filing an issue for the project to properly advertise its licence. Pass this module to the tool via '--ignoreModules ${name}' to suppress this message.`
-        );
+  licenses.forEach(
+    ({
+      name,
+      version,
+      license,
+      licenseText,
+      licenseURLs,
+      path: modulePath,
+    }) => {
+      if (license?.toUpperCase() === "UNLICENSED") {
+        // Ignore unlicensed/private packages
+        return;
       }
-      licenseText = `${license} (${licenseURLs.join(" ")})`;
-    }
 
-    const trimmedText = licenseText.replace(/\r\n|\r|\n/g, EOL).trim();
-    output.push(`${name} ${version}${EOL}--${EOL}${trimmedText}`);
-  });
+      if (!licenseText) {
+        if (!license && (!licenseURLs || licenseURLs.length === 0)) {
+          throw new Error(
+            `No license information found for package '${name}'. Consider filing an issue for the project to properly advertise its licence. Pass this module to the tool via '--ignoreModules ${name}' to suppress this message.`
+          );
+        }
+
+        const copyright = getPackageAuthor(modulePath);
+
+        if (copyright) {
+          licenseText = `Copyright ${copyright}; licensed under ${license} (${licenseURLs.join(" ")})`;
+        } else {
+          licenseText = `Licensed under ${license} (${licenseURLs.join(" ")})`;
+        }
+      }
+
+      const trimmedText = licenseText.replace(/\r\n|\r|\n/g, EOL).trim();
+      output.push(`${name} ${version}${EOL}--${EOL}${trimmedText}`);
+    }
+  );
 
   if (additionalText) {
     output.push(additionalText.join(EOL));
